+++ src/main/java/com/deliver/model/SmartCupboardPosition.java
@@ -12,7 +12,7 @@ public class SmartCupboardPosition {
     @Id
     @GeneratedValue
     @Column(name = "id")
-    private long mId;
+    private int mId;
 
     @Column(name = "empty_full", nullable = false)
     private boolean mEmptyOrFull; // constant.Constant
@@ -31,11 +31,11 @@ public class SmartCupboardPosition {
         this.mColumn = mColumn;
     }
 
-    public long getmId() {
+    public int getmId() {
         return mId;
     }
 
-    public void setmId(long mId) {
+    public void setmId(int mId) {
         this.mId = mId;
     }
